Move shaderLoader plugin. See #301
