+++ source/gloperate/include/gloperate/loaders/ShaderLoader.h
