--- source/plugins/gloperate-loaders/plugin.cpp
@@ -1,9 +1,0 @@
-
-#include <cppexpose/plugin/plugin_api.h>
-
-#include "ShaderLoader.h"
-
-
-CPPEXPOSE_PLUGIN_LIBRARY
-
-CPPEXPOSE_PLUGIN_LIBRARY_END
